Implementing regid functionality
Adding ?regid=example123 to the end of the URL now saves the string to the .csv file as a hidden field.

Useful for keeping track of which presentation the survey response is for. (For example, if the presentation ID is ABC123, change the URL to URL.com?regid=ABC123, and that response will be saved to the .csv along with the presentation ID.)