Form Overhaul
Totally reworked the survey form to have questions based on the original paper Course Evaluation Form.

Still working on handling the .csv formatting a little differently, as there are some quirks with the text fields. Would also like to add a separate text box in the file or another row that also shows the actual questions and not just the question labels (Ex. "Question 1 | How are you?" instead of just "Question1")

Minor style formatting changes.