Previous stars now turn gold
When a star rating is selected, the previous stars now turn gold as well.

There's a small bug when submitting an incomplete form and being sent back to complete it. Only the selected star ratings will remain gold, and the previous stars will not. However, this should be pretty inconsequential and clicking on a new rating fixes the problem anyways.