Fix for Internal Server Error
Added an if/else block to check if the relevant class/registration had an instructor specified in RAMCO.

This fixes an internal server error that would fail to render the survey form if there was no instructor. Now, if no instructor is found, the survey form is rendered without the "Instructor:" text box as well.